block.h
FIX : Structural Tool Analysis
Warnings :
1> 'szx' with type 'unsigned short' 16b is promoted to type 'int' 32b in '(num + 1) << (szx + 4)', then sign-extended to type 'unsigned long' 64b
2> The value of an arithmetic expression (num + 1) << (szx + 4) is subject to overflow due to a failure to cast operands to a larger data type before perfoming arithmetic

+++ resource/csdk/connectivity/lib/libcoap-4.1.1/include/coap/block.h
@@ -73,7 +73,7 @@ unsigned int coap_opt_block_num(const coap_opt_t *block_opt);
  */
 INLINE_API int coap_more_blocks(size_t data_len, unsigned int num, unsigned short szx)
 {
-    return ((num + 1) << (szx + 4)) < data_len;
+    return (((size_t)num + (size_t)1) << ((size_t)szx + (size_t)4)) < data_len;
 }
 
 /** Sets the More-bit in @p block_opt */
